cmake: Move CMake API classes to common
